# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * D Himmatlal and Company API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SiteSettings {
  id: number;
  businessName: string;
  logoText: string;
  /** @nullable */
  logoImageUrl?: string | null;
  phone: string;
  email: string;
  address: string;
  whatsappNumber: string;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  footerText: string;
  createdAt: string;
  updatedAt: string;
}

export interface SiteSettingsInput {
  businessName: string;
  logoText: string;
  /** @nullable */
  logoImageUrl?: string | null;
  phone: string;
  email: string;
  address: string;
  whatsappNumber: string;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  footerText: string;
}

export interface HeroSection {
  id: number;
  heading: string;
  subheading: string;
  primaryButtonText: string;
  primaryButtonLink: string;
  /** @nullable */
  secondaryButtonText?: string | null;
  /** @nullable */
  secondaryButtonLink?: string | null;
  /** @nullable */
  backgroundImageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface HeroSectionInput {
  heading: string;
  subheading: string;
  primaryButtonText: string;
  primaryButtonLink: string;
  /** @nullable */
  secondaryButtonText?: string | null;
  /** @nullable */
  secondaryButtonLink?: string | null;
  /** @nullable */
  backgroundImageUrl?: string | null;
}

export interface AboutSection {
  id: number;
  businessStory: string;
  founderName: string;
  legacyText: string;
  yearsOfExperience: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  founderImageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AboutSectionInput {
  businessStory: string;
  founderName: string;
  legacyText: string;
  yearsOfExperience: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  founderImageUrl?: string | null;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  displayOrder: number;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CategoryInput {
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  displayOrder: number;
  isActive: boolean;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  fullDescription?: string | null;
  /** @nullable */
  price?: string | null;
  contactForPrice: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  galleryImages?: string | null;
  /** @nullable */
  fabric?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
  inStock: boolean;
  isFeatured: boolean;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  name: string;
  slug: string;
  categoryId: number;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  fullDescription?: string | null;
  /** @nullable */
  price?: string | null;
  contactForPrice: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  galleryImages?: string | null;
  /** @nullable */
  fabric?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
  inStock: boolean;
  isFeatured: boolean;
  isActive: boolean;
}

export interface GalleryImage {
  id: number;
  /** @nullable */
  title?: string | null;
  imageUrl: string;
  /** @nullable */
  category?: string | null;
  displayOrder: number;
  createdAt: string;
}

export interface GalleryImageInput {
  /** @nullable */
  title?: string | null;
  imageUrl: string;
  /** @nullable */
  category?: string | null;
  displayOrder: number;
}

export interface Blog {
  id: number;
  title: string;
  slug: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  publishedDate?: string | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface BlogInput {
  title: string;
  slug: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  publishedDate?: string | null;
  isActive: boolean;
}

export interface Testimonial {
  id: number;
  customerName: string;
  review: string;
  rating: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface TestimonialInput {
  customerName: string;
  review: string;
  rating: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
}

export interface ContactDetails {
  id: number;
  address: string;
  phone: string;
  email: string;
  whatsappNumber: string;
  /** @nullable */
  googleMapEmbed?: string | null;
  /** @nullable */
  businessHours?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ContactDetailsInput {
  address: string;
  phone: string;
  email: string;
  whatsappNumber: string;
  /** @nullable */
  googleMapEmbed?: string | null;
  /** @nullable */
  businessHours?: string | null;
}

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  message: string;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  productName?: string | null;
  type: string;
  createdAt: string;
}

export interface InquiryInput {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  message: string;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  productName?: string | null;
  type: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  email: string;
  token: string;
}

export interface DashboardStats {
  totalProducts: number;
  totalCategories: number;
  totalInquiries: number;
  totalBlogs: number;
  totalTestimonials: number;
  totalGalleryImages: number;
  recentInquiries: Inquiry[];
}

export type GetCategoriesParams = {
activeOnly?: boolean;
};

export type GetProductsParams = {
categoryId?: number;
featured?: boolean;
activeOnly?: boolean;
};

export type GetBlogsParams = {
activeOnly?: boolean;
};

export type GetTestimonialsParams = {
activeOnly?: boolean;
};