# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: D Himmatlal and Company API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List all categories
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /gallery:
    get:
      operationId: getGallery
      tags: [gallery]
      summary: List gallery images
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: createGalleryImage
      tags: [gallery]
      summary: Add a gallery image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "201":
          description: Created gallery image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
  /gallery/{id}:
    put:
      operationId: updateGalleryImage
      tags: [gallery]
      summary: Update a gallery image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "200":
          description: Updated gallery image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
    delete:
      operationId: deleteGalleryImage
      tags: [gallery]
      summary: Delete a gallery image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /blogs:
    get:
      operationId: getBlogs
      tags: [blogs]
      summary: List blog posts
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Blog"
    post:
      operationId: createBlog
      tags: [blogs]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogInput"
      responses:
        "201":
          description: Created blog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
  /blogs/{id}:
    get:
      operationId: getBlog
      tags: [blogs]
      summary: Get a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
    put:
      operationId: updateBlog
      tags: [blogs]
      summary: Update a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogInput"
      responses:
        "200":
          description: Updated blog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
    delete:
      operationId: deleteBlog
      tags: [blogs]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /inquiries:
    get:
      operationId: getInquiries
      tags: [inquiries]
      summary: List all inquiries (admin)
      responses:
        "200":
          description: Inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inquiry"
    post:
      operationId: createInquiry
      tags: [inquiries]
      summary: Submit a contact/product inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryInput"
      responses:
        "201":
          description: Inquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
  /inquiries/{id}:
    delete:
      operationId: deleteInquiry
      tags: [inquiries]
      summary: Delete an inquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getAdminMe
      tags: [auth]
      summary: Get current admin session
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        displayOrder: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, name, slug, displayOrder, isActive, createdAt, updatedAt]
    CategoryInput:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        displayOrder: { type: integer }
        isActive: { type: boolean }
      required: [name, slug, displayOrder, isActive]
...
    GalleryImage:
      type: object
      properties:
        id: { type: integer }
        title: { type: ["string", "null"] }
        imageUrl: { type: string }
        category: { type: ["string", "null"] }
        displayOrder: { type: integer }
        createdAt: { type: string }
      required: [id, imageUrl, displayOrder, createdAt]
    GalleryImageInput:
      type: object
      properties:
        title: { type: ["string", "null"] }
        imageUrl: { type: string }
        category: { type: ["string", "null"] }
        displayOrder: { type: integer }
      required: [imageUrl, displayOrder]
    Blog:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        content: { type: string }
        imageUrl: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        publishedDate: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, title, slug, content, isActive, createdAt, updatedAt]
    BlogInput:
      type: object
      properties:
        title: { type: string }
        slug: { type: string }
        content: { type: string }
        imageUrl: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        publishedDate: { type: ["string", "null"] }
        isActive: { type: boolean }
      required: [title, slug, content, isActive]
...
    Inquiry:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        message: { type: string }
        productId: { type: ["integer", "null"] }
        productName: { type: ["string", "null"] }
        type: { type: string }
        createdAt: { type: string }
      required: [id, name, email, message, type, createdAt]
    InquiryInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        message: { type: string }
        productId: { type: ["integer", "null"] }
        productName: { type: ["string", "null"] }
        type: { type: string }
      required: [name, email, message, type]
    LoginInput:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        email: { type: string }
        token: { type: string }
      required: [email, token]
...